+++ shared/src/search/parser/providers.ts
@@ -1,12 +1,12 @@
 import * as Monaco from 'monaco-editor'
-import { Observable, fromEventPattern, of } from 'rxjs'
+import { Observable, fromEventPattern, of, combineLatest } from 'rxjs'
 import { parseSearchQuery } from './parser'
 import { map, first, takeUntil, publishReplay, refCount, switchMap } from 'rxjs/operators'
 import { getMonacoTokens } from './tokens'
 import { getDiagnostics } from './diagnostics'
 import { getCompletionItems } from './completion'
 import { getHoverResult } from './hover'
-import { SearchSuggestion } from '../../graphql/schema'
+import { SearchSuggestion, SearchPatternType } from '../../graphql/schema'
 
 interface SearchFieldProviders {
     tokens: Monaco.languages.TokensProvider
@@ -31,6 +31,7 @@ const alphabet = 'abcdefghijklmnopqrstuvwxyz'
  */
 export function getProviders(
     searchQueries: Observable<string>,
+    patternTypes: Observable<SearchPatternType>,
     fetchSuggestions: (input: string) => Observable<SearchSuggestion[]>
 ): SearchFieldProviders {
     const parsedQueries = searchQueries.pipe(
@@ -81,8 +82,10 @@ export function getProviders(
                     )
                     .toPromise(),
         },
-        diagnostics: parsedQueries.pipe(
-            map(({ parsed }) => (parsed.type === 'success' ? getDiagnostics(parsed.token) : []))
+        diagnostics: combineLatest([parsedQueries, patternTypes]).pipe(
+            map(([{ parsed }, patternType]) =>
+                parsed.type === 'success' ? getDiagnostics(parsed.token, patternType) : []
+            )
         ),
     }
 }